Fix main process in example
